Fix issue #143: Modularize the GitHub issue orchestration runner
